Initial Commit
This Application was written in Android Studio using Java by Charalambos Rotsides as the final project for the course ECE464. This application can track your habits and tasks and give you experience in completing them. This application uses Firebase to store login and personal information such as tasks, experience, level, and habits.